1980 Liberal-Conservative Scales

Liberal-Conservative 7-point scales from the 1980 National Election Study. Includes (in order) self-placement, and rankings of Carter, Reagan, Kennedy, Anderson, Republican party, Democratic Party. Stored as a matrix of integers. The numbers 0, 8, and 9 are considered to be missing values.

Usage

data(LC1980)

Format

An object of class matrix (inherits from array) with 888 rows and 7 columns.

Source

American National Election Studies.

This dataset was originally part of the basicspace package under the same name ("LC1980").

Examples

data(LC1980)
LC1980[LC1980 == 0 | LC1980 == 8 | LC1980 == 9] <- NA
head(LC1980)
#>    Self Carter Reagan Kennedy Anderson Republicans Democrats
#> 1     6      2      6       1        7           5         5
#> 8     6      4      6       4        7           6         4
#> 9     6      3      6       3        3           6         2
#> 10    4      6      4       3        3           5         4
#> 11    5      7      2       5        5           7         5
#> 13    7      6      6       2        5           7         4